% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JKGLM.R
\name{JKGLM}
\alias{JKGLM}
\title{Compute SNP p-values for GWAS by GLM}
\usage{
JKGLM(X, y, CV = NULL, PC = NULL)
}
\arguments{
\item{X}{Matrix containing genotype data organized with SNPs as columns, individuals as rows}

\item{y}{Matrix containing SNP positions and chromosomes corresponding to X}

\item{CV}{Optional argument to include matrix containing covariates. Default is NULL.}

\item{PC}{Optional argument to include matrix containing principal components such as an object of JKPCA. Default is NULL.}
}
\value{
Pvals array of length ncol of X for SNPs from fitted GLM
}
\description{
Compute p-values for the association tests between phenotype and SNPs by Generalized Linear Model fitted for the fixed effects of SNP, with option to include Principal Compoenents and Covariates.
}
