% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JKManhattan.R
\name{JKManhattan}
\alias{JKManhattan}
\title{Manhattan Plot for GWAS Visualization}
\usage{
JKManhattan(Pvals, SNP, sigcutoff = NULL, QTN = NULL)
}
\arguments{
\item{Pvals}{Input vector of Pvals such as an object returned by the JKGLM function}

\item{SNP}{Input matrix containing SNP location. Must include columns Position and Chromosome}

\item{QTN}{Vector of QTN positions that is provided by the user to highlight position in the Manhattan plot. If NULL, then QTN will not be identified.}

\item{sig.cutoff}{Significance threshold for visualization. If NULL, then uses bonferroni correction with alpha = 0.05}
}
\value{
Manhattan plot with user inputs.
}
\description{
Visualize the GWAS by GLM results by Manhattan plot. User can also specify QTN.
}
