% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JKPCA.R
\name{JKPCA}
\alias{JKPCA}
\title{Principle Component Analysis}
\usage{
JKPCA(X, CV = NULL, npc = 5)
}
\arguments{
\item{X}{Markers data in the form n by m with n number of individuals and m number of markers}

\item{CV}{Covariates matrix in the form n by t with n number of individuals and t number of co-variates}

\item{npc}{Number of Principle Components (PCs) that are specified by the user}
}
\value{
Principle Component Analysis
}
\description{
Removes PCs that are linearly dependent with the given covariates and also user can specify how many PCs to chooose as co-factors
}
