% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JKQQ.R
\name{JKQQ}
\alias{JKQQ}
\title{QQ plot}
\usage{
JKQQ(Pvals)
}
\arguments{
\item{Pvals}{Input vector of Pvals such as an object returned by the JKGLM function}
}
\value{
QQplot with user inputs.
}
\description{
Function to generate QQ-plot
}
